requirements.txt
fastapi==0.115.0
starlette==0.38.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-core==2.23.4
typing-extensions==4.12.2
python-dotenv==1.0.1
requests==2.32.3
pymongo==4.8.0
dnspython==2.8.0
openrouteservice==2.3.3
googlemaps==4.10.0
twilio==9.3.6
sendgrid==6.11.0
slowapi==0.1.9
limits==3.13.0
itsdangerous==2.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
python-multipart==0.0.9
email-validator==2.2.0
sqlmodel==0.0.37
psycopg2-binary==2.9.10
certifipymysql
cryptography
